#400 [refactor] writerName 서비스 계층 계층 분리
