+++ module-domain/src/main/java/com/mile/user/service/UserService.java
@@ -1,12 +1,11 @@
 package com.mile.user.service;
 
 import com.mile.client.SocialType;
-import com.mile.client.dto.UserLoginRequest;
 import com.mile.moim.service.dto.MoimListOfUserResponse;
 import com.mile.moim.service.dto.MoimOfUserResponse;
-import com.mile.strategy.LoginStrategyManager;
-import com.mile.strategy.dto.UserInfoResponse;
 import com.mile.user.domain.User;
+import com.mile.writername.service.WriterNameRemover;
+import com.mile.writername.service.WriterNameRetriever;
 import com.mile.writername.service.WriterNameService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,7 +18,8 @@ public class UserService {
     private final UserRetriever userRetriever;
     private final UserRemover userRemover;
     private final UserCreator userCreator;
-    private final WriterNameService writerNameService;
+    private final WriterNameRemover writerNameRemover;
+    private final WriterNameRetriever writerNameRetriever;
 
     public User findById(final Long userId) {
         return userRetriever.findById(userId);
@@ -39,13 +39,13 @@ public class UserService {
 
     public void deleteUser(final Long userId) {
         userRemover.delete(userId);
-        writerNameService.deleteWriterNameByUserId(userId);
+        writerNameRemover.deleteWriterNameByUserId(userId);
     }
 
     public MoimListOfUserResponse getMoimOfUserList(
             final Long userId
     ) {
-        return MoimListOfUserResponse.of(writerNameService.getMoimListOfUser(userId)
+        return MoimListOfUserResponse.of(writerNameRetriever.getMoimListOfUser(userId)
                 .stream()
                 .map(MoimOfUserResponse::of)
                 .collect(Collectors.toList()));
