+++ module-domain/src/main/java/com/mile/writername/service/WriterNameService.java
@@ -1,25 +1,25 @@
 package com.mile.writername.service;
 
-import com.mile.comment.service.CommentGetService;
+import com.mile.comment.service.CommentRemover;
+import com.mile.comment.service.CommentRetriever;
+import com.mile.curious.service.CuriousRemover;
 import com.mile.exception.message.ErrorMessage;
 import com.mile.exception.model.BadRequestException;
-import com.mile.exception.model.ConflictException;
 import com.mile.exception.model.ForbiddenException;
-import com.mile.exception.model.NotFoundException;
 import com.mile.moim.domain.Moim;
+import com.mile.moim.service.MoimRetriever;
 import com.mile.moim.service.dto.MoimWriterNameListGetResponse;
 import com.mile.moim.service.dto.WriterMemberJoinRequest;
-import com.mile.post.domain.Post;
+import com.mile.post.service.PostDeleteService;
 import com.mile.post.service.PostGetService;
 import com.mile.user.domain.User;
+import com.mile.user.service.UserRetriever;
 import com.mile.writername.domain.WriterName;
-import com.mile.writername.repository.WriterNameRepository;
 import com.mile.writername.service.dto.WriterNameDescriptionResponse;
 import com.mile.writername.service.dto.WriterNameDescriptionUpdateRequest;
 import com.mile.writername.service.dto.WriterNameInfoResponse;
 import com.mile.writername.service.dto.WriterNameShortResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -27,43 +27,49 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 
 @Service
 @RequiredArgsConstructor
 public class WriterNameService {
-    private final WriterNameRepository writerNameRepository;
+    private final MoimRetriever moimRetriever;
+    private final UserRetriever userRetriever;
+    private final CommentRemover commentRemover;
+    private final PostDeleteService postDeleteService;
+    private final CuriousRemover curiousRemover;
     private final PostGetService postGetService;
-    private final CommentGetService commentGetService;
-    private static final int MIN_TOTAL_CURIOUS_COUNT = 0;
+    private final CommentRetriever commentGetService;
+
+    private final WriterNameRemover writerNameRemover;
+    private final WriterNameRetriever writerNameRetriever;
+    private final WriterNameCreator writerNameCreator;
+
     private static final int WRITERNAME_PER_PAGE_SIZE = 5;
     private static final int WRITERNAME_MAX_SIZE = 5;
 
-    public WriterName findById(
-            final Long writerNameId
-    ) {
-        return writerNameRepository.findById(writerNameId)
-                .orElseThrow(
-                        () -> new NotFoundException(ErrorMessage.WRITER_NOT_FOUND)
-                );
-    }
-
-    public boolean isUserInMoim(
-            final Long moimId,
-            final Long writerId
-    ) {
-        return writerNameRepository.existsWriterNameByMoimIdAndWriterId(moimId, writerId);
-    }
 
     public WriterNameDescriptionResponse findWriterNameDescription(
             final Long userId,
             final Long writerNameId
     ) {
-        WriterName writerName = findById(writerNameId);
+        WriterName writerName = writerNameRetriever.findById(writerNameId);
         checkWriterNameUserId(userId, writerName);
         return WriterNameDescriptionResponse.of(writerName);
+    }
+
+    public void deleteWriterNameById(
+            final Long writerNameId,
+            final Long userId
+    ) {
+        WriterName writerName = writerNameRetriever.findById(writerNameId);
+        moimRetriever.authenticateOwnerOfMoim(writerName.getMoim(), userRetriever.findById(userId));
+
+        postDeleteService.deleteAllPostByWriterNameId(writerNameId);
+        commentRemover.deleteAllCommentByWriterNameId(writerNameId);
+        curiousRemover.deleteAllByWriterNameId(writerName);
+
+        writerNameRemover.deleteWriterName(writerName);
     }
 
     private void checkWriterNameUserId(final Long userId, final WriterName writerName) {
@@ -76,7 +82,7 @@ public class WriterNameService {
             final Long moimId,
             final Long userId
     ) {
-        return WriterNameShortResponse.of(findByMoimAndUser(moimId, userId));
+        return WriterNameShortResponse.of(writerNameRetriever.findByMoimAndUser(moimId, userId));
     }
 
     @Transactional
@@ -85,7 +91,7 @@ public class WriterNameService {
             final Long writerNameId,
             final WriterNameDescriptionUpdateRequest request
     ) {
-        WriterName writerName = findById(writerNameId);
+        WriterName writerName = writerNameRetriever.findById(writerNameId);
         checkWriterNameUserId(userId, writerName);
         writerName.updateInformation(request);
     }
@@ -93,116 +99,18 @@ public class WriterNameService {
     private void checkWriterNameOverFive(
             final User user
     ) {
-        if (writerNameRepository.countAllByWriter(user) >= WRITERNAME_MAX_SIZE) {
+        if (writerNameRetriever.countWriterNameByWriter(user) >= WRITERNAME_MAX_SIZE) {
             throw new BadRequestException(ErrorMessage.EXCEED_MOIM_MAX_SIZE);
         }
     }
 
-    public void deleteWriterNameByUserId(
-            final Long userId
-    ) {
-        writerNameRepository.delete(findByWriterId(userId));
-    }
 
-    public Long getWriterNameIdByMoimIdAndUserId(
-            final Long moimId,
-            final Long userId
-    ) {
-        return findByMoimAndUser(moimId, userId).getId();
-    }
-
-    public WriterName findByMoimAndUser(
-            final Long moimId,
-            final Long writerId
-    ) {
-        return writerNameRepository.findByMoimIdAndWriterId(moimId, writerId)
-                .orElseThrow(
-                        () -> new ForbiddenException(ErrorMessage.USER_AUTHENTICATE_ERROR)
-                );
-    }
-
-    public Optional<WriterName> findMemberByMoimIdAndWriterId(
-            final Long moimId,
-            final Long writerId
-    ) {
-        return writerNameRepository.findByMoimIdAndWriterId(moimId, writerId);
-    }
-
-
-    public WriterName getWriterNameByPostAndUserId(
-            final Post post,
-            final Long userId
-    ) {
-        return writerNameRepository.findByMoimIdAndWriterId(post.getTopic().getMoim().getId(), userId)
-                .orElseThrow(
-                        () -> new NotFoundException(ErrorMessage.WRITER_NOT_FOUND)
-                );
-    }
-
-    public int findNumbersOfWritersByMoimId(
-            final Long moimId
-    ) {
-        return writerNameRepository.findByMoimId(moimId).size();
-    }
-
-
-    public boolean existWriterNamesByMoimAndName(
-            final Moim moim,
-            final String name
-    ) {
-        return writerNameRepository.existsWriterNameByMoimAndNormalizedName(moim, name);
-    }
-
-    public WriterName findByWriterId(
-            final Long writerId
-    ) {
-        return writerNameRepository.findByWriterId(writerId)
-                .orElseThrow(
-                        () -> new NotFoundException(ErrorMessage.WRITER_NOT_FOUND)
-                );
-    }
-
-    public void decreaseTotalCuriousCountByWriterName(
-            final WriterName writerName
-    ) {
-        writerName.decreaseTotalCuriousCount();
-    }
-
-    public void increaseTotalCuriousCountByWriterName(
-            final WriterName writerName
-    ) {
-        writerName.increaseTotalCuriousCount();
-    }
-
-
-    public WriterName findWriterNameByMoimIdAndUserId(
-            final Long moimId,
-            final Long userId
-    ) {
-        return getById(getWriterNameIdByMoimIdAndUserId(moimId, userId));
-    }
-
-    public List<WriterName> findTop2ByCuriousCount(final Long moimid) {
-        return writerNameRepository.findTop2ByMoimIdAndTotalCuriousCountGreaterThanOrderByTotalCuriousCountDesc(moimid, MIN_TOTAL_CURIOUS_COUNT);
-    }
 
     @Transactional
     public Long createWriterName(final User user, final Moim moim, final WriterMemberJoinRequest joinRequest) {
         checkWriterNameOverFive(user);
-        WriterName writerName;
-        try {
-            writerName = writerNameRepository.save(WriterName.of(moim, joinRequest, user));
-        } catch (DataIntegrityViolationException e) {
-            throw new ConflictException(ErrorMessage.WRITER_NAME_ALREADY_EXIST);
-        }
+        WriterName writerName = writerNameCreator.createWriterName(user, moim, joinRequest);
         return writerName.getId();
-    }
-
-    public WriterName getById(final Long writerNameId) {
-        return writerNameRepository.findById(writerNameId)
-                .orElseThrow(
-                        () -> new NotFoundException(ErrorMessage.WRITER_NOT_FOUND)
-                );
     }
 
 
@@ -212,7 +120,7 @@ public class WriterNameService {
     ) {
         final Long moimId = moim.getId();
         PageRequest pageRequest = PageRequest.of(page - 1, WRITERNAME_PER_PAGE_SIZE, Sort.by(Sort.Direction.DESC, "id"));
-        Page<WriterName> writerNamePage = writerNameRepository.findByMoimIdOrderByIdDesc(moimId, pageRequest);
+        Page<WriterName> writerNamePage = writerNameRetriever.findWriterNameByMoimIdOrderByLatest(moimId, pageRequest);
         List<WriterNameInfoResponse> infoResponses = writerNamePage.getContent()
                 .stream()
                 .map(writerName -> WriterNameInfoResponse.of(writerName.getId(), writerName.getName(),
@@ -223,17 +131,9 @@ public class WriterNameService {
 
         return MoimWriterNameListGetResponse.of(
                 writerNamePage.getTotalPages(),
-                findNumbersOfWritersByMoimId(moimId),
+                writerNameRetriever.findNumbersOfWritersByMoimId(moimId),
                 infoResponses
         );
     }
 
-    public List<Moim> getMoimListOfUser(
-            final Long userId
-    ) {
-        return writerNameRepository.findAllByWriterId(userId)
-                .stream()
-                .map(writerName -> writerName.getMoim())
-                .collect(Collectors.toList());
-    }
 }
